requirements.txt
# Core Dependencies
streamlit==1.32.0
pandas==2.2.0
numpy==1.26.3
python-dotenv==1.0.0

# LLM & AI
openai==1.102.0
anthropic==0.64.0
google-generativeai==0.3.2
protobuf>=3.20,<5
langchain==0.1.6
langchain-openai==0.0.5
langchain-anthropic==0.1.1
tiktoken==0.5.2
transformers==4.37.2

# Data Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
beautifulsoup4==4.12.3
requests==2.31.0
yfinance==0.2.36

# Analysis & Visualization
plotly==5.19.0
matplotlib==3.8.2
seaborn==0.13.2
wordcloud==1.9.3
textblob==0.17.1

# NLP & Text Processing
nltk==3.8.1
spacy==3.7.2
scikit-learn==1.4.0

# Testing & Development
pytest==8.0.0
pytest-cov==4.1.0
black==24.1.1
flake8==7.0.0
ipykernel==6.29.0
jupyter==1.0.0

# Utilities
tqdm==4.66.1
colorama==0.4.6
python-dateutil==2.8.2
pytz==2024.1